[feat] js 파일 경로 및 일부 html 수정 :memo:
