+++ client/cart.html
@@ -10,8 +10,7 @@
     />
     <link rel="stylesheet" href="./css/style.css" />
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
-    <script src="./js/main/index.js" type="module"></script>
-    <script src="./js/Cart/index.js" type="module"></script>
+    <script src="./js/Cart/cart.js" type="module"></script>
     <title>CART</title>
   </head>
   <body>
@@ -75,8 +74,8 @@
         </div>
 
         <ul class="header--member">
-          <li><a href="./register.html" class="is-active">회원가입</a></li>
-          <li>
+          <li class="register"><a href="./register.html" class="is-active">회원가입</a></li>
+          <li class="signIn">
             <a href="./login.html">로그인</a>
           </li>
           <li>
