+++ client/index.html
@@ -12,7 +12,8 @@
     <link rel="stylesheet" href="./css/style.css" />
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
 
-    <script src="./js/index.js" type="module"></script>
+    <script defer src="./js/main/index.js" type="module"></script>
+    
     <title>마켓칼리</title>
   </head>
   <body>
