--- client/js/Cart/index.js
@@ -1,1 +1,0 @@
-export * from "./cart.js";
