--- client/js/index.js
@@ -1,5 +1,0 @@
-export * from "./main/index.js";
-export * from "./page/index.js";
-export * from "./Cart/index.js";
-export * from "./Product/index.js";
-export * from "./Sign/index.js";
