--- client/js/page/index.js
@@ -1,2 +1,0 @@
-export * from "./register.js";
-export * from "./login.js";
