+++ client/login.html
@@ -7,8 +7,8 @@
     <title>로그인</title>
     <link rel="stylesheet" href="./css/pages/login.css" />
     <link rel="stylesheet" href="./css/style.css" />
-    <script src="./js/index.js" type="module"></script>
-    <script src="./js/dropDown.js" type="module"></script>
+    <script src="./js/page/login.js" type="module"></script>
+    <script src="./js/main/dropDown.js" type="module"></script>
   </head>
   <body>
     <!-- Top banner start -->
