+++ client/product_list.html
@@ -9,7 +9,6 @@
       href="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.css"
     />
     <link rel="stylesheet" href="./css/style.css" />
-
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
     <script src="./js/Product/index.js" type="module"></script>
     <script src="./js/main/dropDown.js" type="module"></script>
@@ -77,8 +76,8 @@
         </div>
 
         <ul class="header--member">
-          <li><a href="./register.html" class="is-active">회원가입</a></li>
-          <li>
+          <li class="register"><a href="./register.html" class="is-active">회원가입</a></li>
+          <li class="signIn">
             <a href="./login.html">로그인</a>
           </li>
           <li>
