+++ client/register.html
@@ -8,8 +8,8 @@
     <title>회원가입</title>
     <link rel="stylesheet" href="./css/pages/register.css" />
     <link rel="stylesheet" href="./css/style.css" />
-    <script src="./js/main/dropDown.js" type="module"></script>
     <script src="./js/page/register.js" type="module"></script>
+    <script src="./js/main/dropDown.js" type="module"></script>
     <!-- 주소 검색 API -->
     <script src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"></script>
   </head>
